1. It shows how to create a decorator function that accesses a function that's passed in as a property without creating its own isolated scope.
2. Usage of parse service to get the function pointer of the "handlePause(evt)" method defined in the html as value of event-pause attribute
3.  it's very important whenever you have an event that fires that Angular doesn't know about, such as an HTML elements event, that you start a digest cycle by calling scope.apply.
The HTML5 video has fired the pause event, but that's actually happening outside of Angular's knowledge. So if we were to try to do anything with Angular, in Angular nothing happens outside of a _____ cycle. So we're going to have to start a new digest cycle, because an external event has happened. And we do that by calling scope.apply, and that function is $apply.

fn(scope, { evt: event }) -  here evt: is like function parameter overloading - i.e. we are passing new values to this evt: parameter.