Added Search Component for allClassrooms.component.html and used query string in webapi
1. Created a new classroomsearch.component.ts file
here, defined a Subject<T> type observable

A Subject is a producer of an observable event stream; Each call to search puts a new string into this subject's observable stream by calling next.
Using this Subject type, a search is affected on every keystroke which may create too much server request.
So, to limit this,
debounceTime, distinctUntilChanged and switchMap are defined.
debounceTime(300)        // wait for 300ms pause in events
            .distinctUntilChanged()   // ignore if next search term is same as previous
            .switchMap(term => term   // switch to new observable each time i.e new array is created for each search

2. search method => creates a new search on keystroke

3. gotoDetail => this method navigates the user to selected classroom

4. used async keyword = >the async pipe (AsyncPipe) subscribes to the Observable and produces the array of classroom to *ngFor

5. Used [FromQuery] to read the query string parameter for classroom webapi get request to return the search results