Added activitysearch feature using routing with query param, filter pipe transform, resolve service and nesting components
1. Created a new activitysearch.component and used it in the allActivities.component.html
this activitysearch.component -  search for the activity for a classroom in particular month
and display the details in classroomDetail.component.html

2. activitysearch.component creates the navigation route using router.navigate mthod
Here, 2 query parameters are passed => id and month
this.router.navigate(['/classroomdetail', { id: id, month: month }])
the url is like
http://localhost/#/classroomdetail;id=1;month=1
The optional route parameters are not separated by "?" and "&" as they would be in the URL query string. They are separated by semicolons ";"
This is matrix URL notation.

3. Created a filteredActivities.pipe to filer the result set
used this filer in classroomDetail.component.html
*ngFor="let act of (classroom.activity | filteredActivities:month)"
Here, filteredActivities => name of Pipe
classroom.activity => first parameter to Pipe
month is second parameter to Pipe

4. Added a classroomDetail.resolve => to ensure that data is loded before the classroomDetail.component.html is displayed

5. In schoolapp-router.module => path: 'classroomdetail' without any parameters to accept matrix URL notation