Created a Form Array to dynamicall add controls in the form
1. Defined a controls as formarray and added a formgroup into it
formgroup consist of address related controls
2. Created a Getter property to access the addresses form array in the html file
get addresses():FormArray {}

This get property is very important as without it the form array can not be referenced in the html file

3. Defined the formArrayName as the addresses get property to set a div as form array
and used ngFor loop to iterate over each control in the form group.

here, one important thing is that the formgroup name is equal to the index location
in the form array.
[formGroupName]="i"

4. Also, created an Add Address button to dynamically add another address form group in the form array whenever the button is clicked.