Added code to CreateToken using JWT - Jason Web Token
To create a JWT token few steps are required:
1. JwtSecurityToken requires =>  issuer: , audience:,  claims: expires: , signingCredentials values as defined below:

var token = new JwtSecurityToken(
                          issuer: _config["Tokens:Issuer"],
                          audience: _config["Tokens:Audience"],
                          claims: claims,
                          expires: DateTime.UtcNow.AddMinutes(15),
                          signingCredentials: creds
                          );

2. Here, key is created using a secret known only to the application.
3. signingCredentials  => is created using SigningCredentials class provided by JWT package.
Once, the token is created, we need to add the JWT middleware as well in startup.cs file.

4. JWT middleware is added using => UseJwtBearerAuthentication
This middleware requires => JwtBearerOptions
which further needs = > TokenValidationParameters

Setting up all this ensures that the JWT token is generated and sent beck to client.
The generated token can be verified in https://jwt.io/. This site breaks up the token in 3 differnet pieces.
1. Header, 2. Payload, 3 Signature or seceret key