Added support for Post request for AuthorsController  and BooksApiController
This commit has support for creation of
1. Authors
2. Book for a specific Author
I have created separate model for Post request so as to keep the separation of concerns between GET and POST requests
There may be some validation that I add in Post request and some calculated data that I display in GET request.
So, to avoid such situation, its good practice to create different classes for GET and POST request.