Setting the document or window title using the Title service.
The problem with <title>

The obvious approach is to bind a property of the component to the HTML <title> like this:

<title>{{This_Does_Not_Work}}</title>
Sorry but that won't work. The root component of our application is an element contained within the <body> tag. The HTML <title> is in the document <head>, outside the body, making it inaccessible to Angular data binding.

We could grab the browser document object and set the title manually. That's dirty and undermines our chances of running the app outside of a browser someday.

Use the Title service

Fortunately, Angular bridges the gap by providing a Title service as part of the Browser platform. The Title service is a simple class that provides an API for getting and setting the current HTML document title:

getTitle() : string — Gets the title of the current HTML document.
setTitle( newTitle : string ) — Sets the title of the current HTML document.